% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprr.R
\name{init_packages}
\alias{init_packages}
\title{init_packages}
\usage{
init_packages(pkgs)
}
\arguments{
\item{pkgs}{character vector of the packages used in your project}
}
\value{
nothing
}
\description{
init_packages
}
\examples{
init_packages(c("crosstable", "dplyr", "officer"))
}
